4.23 fix tests

+++ round4/bloglist/tests/blog_api.test.js
@@ -1,9 +1,32 @@
 const supertest = require('supertest');
+const jwt = require('jsonwebtoken');
 const app = require('../src/app');
 const blog = require('../src/models/blog');
+const user = require('../src/models/user');
 const db = require('../src/utils/db');
+const config = require('../src/utils/config');
 
 const api = supertest(app);
+
+let testUser;
+let authToken;
+beforeEach(async () => {
+  await db.connect();
+  await blog.deleteAll();
+  await user.deleteAll();
+  await user.create({
+    username: 'blogTest',
+    name: 'test',
+    password: 'test',
+  });
+  testUser = await user.getByUsername('blogTest');
+  authToken = jwt.sign({
+    username: testUser.username,
+    id: testUser._id,
+  }, config.LOGIN_SECRET);
+
+  await db.close();
+});
 
 describe('GET /api/blogs', () => {
   const testBlogs = [{
@@ -22,13 +45,13 @@ describe('GET /api/blogs', () => {
   beforeEach(async () => {
     await db.connect();
     await blog.deleteAll();
-    await Promise.all(testBlogs.map((testBlog) => blog.create(testBlog)));
+    await blog.create(testBlogs[0], testUser._id);
+    await blog.create(testBlogs[1], testUser._id);
   });
 
   afterEach(async () => {
     await db.close();
   });
-
   test('blogs are returned as json', async () => {
     await api
       .get('/api/blogs')
@@ -45,7 +68,7 @@ describe('GET /api/blogs', () => {
   });
 });
 
-describe('GET /api/blogs', () => {
+describe('POST /api/blogs', () => {
   beforeEach(async () => {
     await db.connect();
     await blog.deleteAll();
@@ -55,14 +78,24 @@ describe('GET /api/blogs', () => {
     await db.close();
   });
 
-  test('blog should have id', async () => {
+  test('should fail with no authentication header', async () => {
+    const testBlog = {
+      title: 'noAuthTest',
+      author: 'postTestAuthor',
+      url: 'test',
+      likes: 1,
+    };
+    await api.post('/api/blogs').send(testBlog).expect(401);
+  });
+
+  test('should work', async () => {
     const testBlog = {
       title: 'postTest',
       author: 'postTestAuthor',
       url: 'test',
       likes: 1,
     };
-    await api.post('/api/blogs').send(testBlog);
+    await api.post('/api/blogs').set('Authorization', `Bearer ${authToken}`).send(testBlog).expect(200);
     const after = await blog.getAll();
     expect(after.length).toBe(1);
     expect(after[0].title).toBe(testBlog.title);
@@ -76,7 +109,7 @@ describe('GET /api/blogs', () => {
       author: 'postTestAuthor',
       url: 'test',
     };
-    await api.post('/api/blogs').send(likesMissing);
+    await api.post('/api/blogs').set('Authorization', `Bearer ${authToken}`).send(likesMissing);
     const after = await blog.getAll();
     expect(after.length).toBe(1);
     expect(after[0].likes).toBe(0);
@@ -87,7 +120,7 @@ describe('GET /api/blogs', () => {
       url: 'test',
       likes: 1,
     };
-    await api.post('/api/blogs').send(titleMissing).expect(400);
+    await api.post('/api/blogs').set('Authorization', `Bearer ${authToken}`).send(titleMissing).expect(400);
   });
   test('missing url should give error 400', async () => {
     const urlMissing = {
@@ -95,7 +128,7 @@ describe('GET /api/blogs', () => {
       author: 'postTestAuthor',
       likes: 1,
     };
-    await api.post('/api/blogs').send(urlMissing).expect(400);
+    await api.post('/api/blogs').set('Authorization', `Bearer ${authToken}`).send(urlMissing).expect(400);
   });
 });
 
@@ -115,25 +148,31 @@ describe('DELETE /api/blogs', () => {
   beforeEach(async () => {
     await db.connect();
     await blog.deleteAll();
-    await blog.create(testBlog);
-    await blog.create(toBeDeleted);
+    await blog.create(testBlog, testUser._id);
+    await blog.create(toBeDeleted, testUser._id);
   });
 
   afterEach(async () => {
     await db.close();
   });
 
+  test('should return 401 with no auth field', async () => {
+    const before = await blog.getAll();
+    expect(before.length).toBe(2);
+    const { id } = before.find((b) => b.title === toBeDeleted.title);
+    await api.delete(`/api/blogs/${id}`).expect(401);
+  });
   test('should work', async () => {
     const before = await blog.getAll();
     expect(before.length).toBe(2);
     const { id } = before.find((b) => b.title === toBeDeleted.title);
-    await api.delete(`/api/blogs/${id}`);
+    await api.delete(`/api/blogs/${id}`).set('Authorization', `Bearer ${authToken}`).expect(200);
     const after = await blog.getAll();
     expect(after.length).toBe(1);
     expect(after.find((b) => b.title === toBeDeleted.title)).toBe(undefined);
   });
   test('should return 400 on malformed id', async () => {
-    await api.delete('/api/blogs/test').expect(400);
+    await api.delete('/api/blogs/test').set('Authorization', `Bearer ${authToken}`).expect(400);
   });
 });
 
@@ -147,7 +186,7 @@ describe('PUT /api/blogs', () => {
   beforeEach(async () => {
     await db.connect();
     await blog.deleteAll();
-    await blog.create(testBlog);
+    await blog.create(testBlog, testUser._id);
   });
 
   afterEach(async () => {
@@ -160,13 +199,13 @@ describe('PUT /api/blogs', () => {
     const { id } = before[0];
     const newLikes = 15;
     const newUrl = 'TEST URL';
-    await api.put(`/api/blogs/${id}`).send({ ...testBlog, likes: newLikes, url: newUrl });
+    await api.put(`/api/blogs/${id}`).send({ ...testBlog, likes: newLikes, url: newUrl }).expect(200);
 
     const after = await blog.getAll();
     expect(after[0].likes).toBe(newLikes);
     expect(after[0].url).toBe(newUrl);
   });
   test('should return 400 on malformed id', async () => {
-    await api.delete('/api/blogs/test').expect(400);
+    await api.put('/api/blogs/test').send(testBlog).expect(400);
   });
 });
